feat: introduce `rb_bundle_fetch()` repository rule
This repository rule deprecates existing `rb_bundle()` and provides a
number of benefits over existing implementation:

1. Hermeticity of `bundle install` - it's run as a regular rule.
2. Ruby toolchain no longer needs to be installed during WORKSPACE
   loading.
2. Downloading of gems is taken care of by Bazel. This also means that
   gems are stored in a repository cache.
3. RBE is potentially supported (at least with JRuby).

The new rule however is not complete yet and lacks few important
features that will be added later.

1. Support for gems installed from Git repositories.
2. Support for checksums introduced in Bundler 2.50.

For most cases, the migration from `rb_bundle()` to `rb_bundle_fetch()`
is straightforward - use the same parameters with an extra `gemfile_lock`.
The main breaking change is that your targets that used to depend on
Bundler binstubs a new `bin` package. For example:

| rb_bundle()         | rb_bundle_fetch    |
|---------------------|--------------------|
| @bundle//:bin/rspec | @bundle//bin:rspec |