rb_binary: include only wrapper script in default outputs
without this, it is not possible to reliably locate (via runfiles
library or make variables) the executable output of a rb_binary target
from another target using it as a data dependency, e.g. a shell script.
adds an example of such a case to the 'gem' example workspace.

additionally try to avoid converting depsets to lists when possible